fix: remove ratelimited endpoints
