+++ src/endpoints/quests/avnu/mod.rs
@@ -1,3 +1,3 @@
 pub mod discord_fw_callback;
-pub mod twitter_fw_callback;
+pub mod verify_twitter_fw;
 pub mod verify_twitter_rt;
