+++ src/endpoints/quests/sithswap/mod.rs
@@ -1,4 +1,4 @@
-pub mod twitter_fw_callback;
+pub mod verify_twitter_fw;
 pub mod verify_added_liquidity;
 pub mod verify_has_root_domain;
 pub mod verify_twitter_rt;
