+++ src/endpoints/quests/starknetid/mod.rs
@@ -1,4 +1,4 @@
-pub mod verify_twitter_follow;
+pub mod verify_twitter_fw;
 pub mod verify_has_domain;
 pub mod verify_has_root_domain;
 pub mod verify_socials;
